[Chore] delete Test file
