--- src/pages/RollingPage/Test2.jsx
@@ -1,11 +1,0 @@
-//이페이지는 /post/{id}/message 페이지임
-
-function Test2() {
-  return (
-    <div>
-      <div>/post/id/message 페이지</div>
-    </div>
-  );
-}
-
-export default Test2;
