+++ src/router.jsx
@@ -2,11 +2,9 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import LandingPage from "./pages/LandingPage/LandingPage";
 import ListPage from "./pages/ListPage/ListPage";
 import ToPage from "./pages/ToPage/ToPage";
-// import RollingPage from "./pages/RollingPage/RollingPage";
+import RollingPage from "./pages/RollingPage/RollingPage";
 import FromPage from "./pages/FromPage/FromPage";
 import Layout from "./styles/Layout";
-import Test from "./pages/RollingPage/Test";
-import Test2 from "./pages/RollingPage/Test2";
 export default function Routers() {
   return (
     <BrowserRouter>
@@ -15,9 +13,7 @@ export default function Routers() {
           <Route index element={<LandingPage />}></Route>
           <Route path="list" element={<ListPage />}></Route>
           <Route path="post" element={<ToPage />}></Route>
-          {/* <Route path="post/:id" element={<RollingPage />}></Route> */}
-          <Route path="post/:id" element={<Test />}></Route>
-          <Route path="post/9727/message" element={<Test2 />}></Route>
+          <Route path="post/:id" element={<RollingPage />}></Route>
           <Route path="post/:id/message" element={<FromPage />}></Route>
         </Route>
       </Routes>
